requirements.txt
appdirs==1.4.4
appnope==0.1.0
backcall==0.2.0
decorator==4.4.2
distlib==0.3.1
Django==2.2.16
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-extensions==3.0.9
filelock==3.0.12
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
Pygments==2.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==5.0.5
virtualenv==20.0.35
wcwidth==0.2.5
